Changes to edit profile and theme style Style can be improved greatly in future
