+++ PosseUpApp/app/src/main/res/layout/activity_main.xml
@@ -15,6 +15,7 @@
             android:layout_height="wrap_content"
             android:layout_width="match_parent"
             android:elevation="5dp"
+            app:theme="@style/RevisedAppTheme"
             android:minHeight="?attr/actionBarSize"
             android:background="@color/ColorPrimary" />
     <FrameLayout
@@ -70,7 +71,7 @@
             app:menu="@menu/menu_drawer"
             app:itemTextColor="@color/drawer_selector"
             app:itemIconTint="@color/drawer_selector"
-            android:background="@color/white">
+            android:background="@android:color/white">
 
         </android.support.design.widget.NavigationView>
