+++ PosseUpApp/app/src/main/res/layout/configure_account_layout.xml
@@ -18,7 +18,7 @@
             android:layout_height="wrap_content"
             android:text="@string/hi"
             android:id="@+id/configure_welcome"
-            android:textColor="@color/white"
+            android:textColor="@color/offWhite"
             android:layout_marginTop="38dp"
             android:layout_alignParentTop="true"
             android:layout_alignStart="@+id/linearLayout" />
@@ -28,7 +28,7 @@
             android:layout_height="wrap_content"
             android:text=""
             android:id="@+id/welcome_name"
-            android:textColor="@color/white"
+            android:textColor="@color/offWhite"
             android:layout_marginTop="38dp"
             android:layout_alignBottom="@+id/configure_welcome"
             android:layout_centerHorizontal="true" />
@@ -43,8 +43,8 @@
         android:ems="10"
         android:id="@+id/fullname"
         android:textSize="12dp"
-        android:textColor="@color/white"
-        android:textColorHint="@color/white"
+        android:textColor="@color/offWhite"
+        android:textColorHint="@color/offWhite"
         android:minWidth="200dp"
         android:layout_centerVertical="true"
         android:layout_centerHorizontal="true" />
@@ -57,8 +57,8 @@
         android:textSize="12dp"
         android:id="@+id/email_signup"
         android:layout_below="@+id/fullname"
-        android:textColor="@color/white"
-        android:textColorHint="@color/white"
+        android:textColor="@color/offWhite"
+        android:textColorHint="@color/offWhite"
         android:minWidth="200dp"
         android:layout_alignStart="@+id/fullname"
         android:hint="Email"
@@ -72,8 +72,8 @@
         android:layout_marginTop="27dp"
         android:hint="@string/password_field"
         android:textSize="12dp"
-        android:textColor="@color/white"
-        android:textColorHint="@color/white"
+        android:textColor="@color/offWhite"
+        android:textColorHint="@color/offWhite"
         android:inputType="textPassword"
         android:fontFamily="sans-serif"
         android:minWidth="200dp"
