+++ PosseUpApp/app/src/main/res/layout/edit_profile_layout.xml
@@ -1,0 +1,77 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:orientation="vertical">
+    <android.support.v7.widget.Toolbar
+        android:id="@+id/edit_profile_activity_toolbar"
+        android:layout_height="wrap_content"
+        android:layout_width="match_parent"
+        android:elevation="5dp"
+        android:minHeight="?attr/actionBarSize"
+        android:background="@color/ColorPrimary" />
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="@dimen/margin_medium"
+        android:layout_marginEnd="@dimen/margin_medium"
+        android:layout_marginTop="@dimen/margin_medium"
+        android:orientation="vertical">
+        <RelativeLayout
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginBottom="@dimen/margin_small">
+            <com.mikhaellopez.circularimageview.CircularImageView
+                android:layout_width="120dp"
+                android:layout_height="120dp"
+                android:src="@drawable/profiler"
+                android:id="@+id/view2" />
+            <android.support.design.widget.FloatingActionButton
+                android:id="@+id/addImage_Button"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+
+                android:layout_gravity="bottom"
+                app:backgroundTint="@color/primaryThemeColor"
+                android:src="@drawable/ic_add"
+                android:clickable="true"
+                app:fabSize="mini"
+                android:layout_alignParentTop="true"
+                android:layout_alignEnd="@+id/view2" />
+        </RelativeLayout>
+        <com.rengwuxian.materialedittext.MaterialEditText
+            android:id="@+id/profile_username_edit"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:hint="Name"
+            app:met_floatingLabel="highlight"
+            app:met_primaryColor="@color/ColorPrimary"
+            android:textColor="@color/offWhite"
+            android:textColorHint="@color/offWhite"
+            android:layout_marginBottom="@dimen/margin_small"
+            android:text="@string/username"
+            android:textSize="@dimen/abc_text_size_small_material" />
+        <com.rengwuxian.materialedittext.MaterialEditText
+            android:id="@+id/profile_bio_edit"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:textColor="@color/offWhite"
+            android:textColorHint="@color/offWhite"
+            android:hint="Bio"
+            app:met_floatingLabel="highlight"
+            app:met_primaryColor="@color/ColorPrimary"
+            android:text="Bio"
+            android:textSize="@dimen/abc_text_size_small_material" />
+        <android.support.v7.widget.AppCompatButton
+            android:id="@+id/change_password_button"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="Change password"
+            android:textColor="@color/offWhite"/>
+
+        <!--TODO maybe add external login stuff here-->
+
+    </LinearLayout>
+
+</LinearLayout>
