+++ PosseUpApp/app/src/main/res/layout/forgotpasswordfrag_layout.xml
@@ -11,8 +11,8 @@
         android:ems="10"
         android:id="@+id/forgotpassword_email"
         android:textSize="12dp"
-        android:textColor="@color/white"
-        android:textColorHint="@color/white"
+        android:textColor="@color/offWhite"
+        android:textColorHint="@color/offWhite"
         android:minWidth="200dp"
         android:layout_below="@+id/signin_desc"
         android:layout_centerHorizontal="true"
@@ -28,8 +28,8 @@
         android:layout_marginTop="19dp"
         android:hint="@string/password_code"
         android:textSize="12dp"
-        android:textColor="@color/white"
-        android:textColorHint="@color/white"
+        android:textColor="@color/offWhite"
+        android:textColorHint="@color/offWhite"
         android:inputType="textCapCharacters"
         android:fontFamily="sans-serif"
         android:minWidth="200dp" />
@@ -47,7 +47,7 @@
         android:paddingRight="35dp"
         android:textAlignment="center"
         android:gravity="center"
-        android:textColor="@color/white" />
+        android:textColor="@color/offWhite" />
 
     <LinearLayout
         android:layout_width="match_parent"
