+++ PosseUpApp/app/src/main/res/layout/profile_edit_password_dialog.xml
@@ -11,6 +11,11 @@
     android:paddingLeft="@dimen/margin_small"
     android:paddingRight="@dimen/margin_small"
     >
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:textSize="@dimen/abc_text_size_body_1_material"
+        android:text="If successful, this will immediately update your password"/>
 
 
 
@@ -19,6 +24,8 @@
         android:layout_height="wrap_content"
         android:inputType="textPassword"
         android:maxLength="32"
+        android:textColor="@color/offWhite"
+        android:textColorHint="@color/offWhite"
         android:hint="Old password"
         app:met_floatingLabel="highlight"
         app:met_primaryColor="@color/ColorPrimary"
@@ -32,6 +39,8 @@
         android:inputType="textPassword"
         android:hint="New password"
         app:met_floatingLabel="highlight"
+        android:textColor="@color/offWhite"
+        android:textColorHint="@color/offWhite"
         android:maxLength="32"
         app:met_primaryColor="@color/ColorPrimary"
         android:fontFamily="sans-serif"
