+++ PosseUpApp/app/src/main/res/layout/register_layout.xml
@@ -51,8 +51,8 @@
         android:ems="10"
         android:id="@+id/username_register"
         android:textSize="12dp"
-        android:textColor="@color/white"
-        android:textColorHint="@color/white"
+        android:textColor="@color/offWhite"
+        android:textColorHint="@color/offWhite"
         android:minWidth="200dp"
         android:layout_below="@+id/view"
         android:layout_centerHorizontal="true"
@@ -66,8 +66,8 @@
         android:textSize="12dp"
         android:id="@+id/email_signup"
         android:layout_below="@+id/username_register"
-        android:textColor="@color/white"
-        android:textColorHint="@color/white"
+        android:textColor="@color/offWhite"
+        android:textColorHint="@color/offWhite"
         android:minWidth="200dp"
         android:layout_alignStart="@+id/username_register"
         android:hint="Email"
@@ -81,8 +81,8 @@
         android:layout_marginTop="22dp"
         android:hint="@string/password_field"
         android:textSize="12dp"
-        android:textColor="@color/white"
-        android:textColorHint="@color/white"
+        android:textColor="@color/offWhite"
+        android:textColorHint="@color/offWhite"
         android:inputType="textPassword"
         android:fontFamily="sans-serif"
         android:minWidth="200dp"
