+++ PosseUpApp/app/src/main/res/layout/signinfrag_layout.xml
@@ -11,8 +11,8 @@
         android:ems="10"
         android:id="@+id/username_signin"
         android:textSize="12dp"
-        android:textColor="@color/white"
-        android:textColorHint="@color/white"
+        android:textColor="@color/offWhite"
+        android:textColorHint="@color/offWhite"
         android:minWidth="200dp"
         android:layout_below="@+id/signin_desc"
         android:layout_centerHorizontal="true"
@@ -28,8 +28,8 @@
         android:layout_marginTop="19dp"
         android:hint="@string/password_field"
         android:textSize="12dp"
-        android:textColor="@color/white"
-        android:textColorHint="@color/white"
+        android:textColor="@color/offWhite"
+        android:textColorHint="@color/offWhite"
         android:inputType="textPassword"
         android:fontFamily="sans-serif"
         android:minWidth="200dp" />
