+++ PosseUpApp/app/src/main/res/values/color.xml
@@ -4,7 +4,7 @@
     <color name="ColorPrimaryDark">#c13257</color>
     <color name="tabsScrollColor">#8a140e</color>
 
-    <color name="white">#FFF8F8</color>
+    <color name="offWhite">#FFF8F8</color>
     <color name="primaryThemeColor">#FF3366</color>
 
     <color name="grey">#757575</color>
